Get rid of internal modules
