+++ packages/expo-web-browser/android/src/main/java/expo/modules/webbrowser/WebBrowserExceptions.kt
@@ -2,9 +2,11 @@ package expo.modules.webbrowser
 
 import expo.modules.kotlin.exception.CodedException
 
-internal class NoPreferredPackageFound(
-  message: String?
-) : CodedException(code = "PREFERRED_PACKAGE_NOT_FOUND", message, cause = null)
+internal class NoPreferredPackageFound : CodedException(
+  code = "PREFERRED_PACKAGE_NOT_FOUND",
+  message = "Cannot determine preferred package without satisfying it",
+  cause = null
+)
 
 internal class PackageManagerNotFoundException : CodedException("Package Manager not found")
