+++ packages/expo-web-browser/android/src/main/java/expo/modules/webbrowser/WebBrowserPackage.kt
@@ -6,10 +6,6 @@ import expo.modules.core.interfaces.InternalModule
 import expo.modules.core.ExportedModule
 
 class WebBrowserPackage : BasePackage() {
-  override fun createInternalModules(context: Context): List<InternalModule> = listOf(
-    InternalCustomTabsActivitiesHelper(),
-    InternalCustomTabsConnectionHelper(context)
-  )
-
+  override fun createInternalModules(context: Context): List<InternalModule> = emptyList()
   override fun createExportedModules(context: Context): List<ExportedModule> = emptyList()
 }
